refactor: move useAssessment to its file
